+++ stories/user-action/button/IconButton.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
-import { ButtonColorUtil, IconButton } from '../../../src'
+import { IconButton, IconButtonUtil } from '../../../src'
 import { action } from 'storybook/actions'
 import { Plus } from 'lucide-react'
 
@@ -9,7 +9,7 @@ const meta = {
   argTypes: {
     color: {
       control: 'select',
-      options: ButtonColorUtil.solid,
+      options: IconButtonUtil.icon,
     }
   },
 } satisfies Meta<typeof IconButton>
@@ -20,10 +20,10 @@ type Story = StoryObj<typeof meta>;
 
 export const iconButton: Story = {
   args: {
-    children: (<Plus className="w-full h-full"/>),
     color: 'negative',
     size: 'medium',
     disabled: false,
     onClick: action('Clicked'),
+    children: (<Plus className="w-full h-full"/>),
   },
 }